requirements.txt
fastapi
uvicorn
pdfplumber
camelot-py[cv]
python-multipart
pandas
numpy
opencv-python
PyPDF2